+++ src/sumCircuit/SumCircuit.kt
@@ -2,22 +2,22 @@ package sumCircuit
 
 import perceptron.NandPerceptron
 
-class SumCircuit(private val x: Int, private val y: Int) {
-    private val calculatedSum: Int by lazy {
+class SumCircuit(private val x: Double, private val y: Double) {
+    private val calculatedSum: Double by lazy {
         val nandPerceptron = NandPerceptron()
         nandPerceptron.calculateOutput(nandPerceptron.calculateOutput(x, nandPerceptron.calculateOutput(x, y)),
                                        nandPerceptron.calculateOutput(nandPerceptron.calculateOutput(x, y), y))
     }
-    private val calculatedCarryBit: Int by lazy {
+    private val calculatedCarryBit: Double by lazy {
         val nandPerceptron = NandPerceptron()
         nandPerceptron.calculateOutput(nandPerceptron.calculateOutput(x, y), nandPerceptron.calculateOutput(x, y))
     }
 
-    fun getSumResult(): Int {
+    fun getSumResult(): Double {
         return this.calculatedSum
     }
 
-    fun getCarryBit(): Int {
+    fun getCarryBit(): Double {
         return this.calculatedCarryBit
     }
 }
